Add overbooking quota driver
Add a new quota_driver: OverbookingDbQuotaDriver, that's
largely based on the existing DbQuotaDriver.

The main difference is that it not only checks project quota like
the original driver, but also checks if aggregate usage of projects
within a domain does not exceed domain's quota.

Driver could be enabled by adding this to the cinder.conf:

  quota_driver = cinder.quota.OverbookingDbQuotaDriver

Change-Id: Iaca7fb0311129f5e8316753279caae4cb378af11